Switch from debian-installer to calamares.
This also makes it easier to switch between stable, stable+backports, testing, and unstable.

Also included are some customizations to calamares-settings-debian, it's cleaner here and
easier to keep in sync with calamares-settings-debian if it's not a full fork.

 * displaymanager.conf: Modified to only run on lightdm, don't reset the default user session.
 * packages.conf: Add calamares and gparted to the list of packages to remove post-install.
 * welcome.conf: Lower disk space needed from 10G to 3G, this is mostly useful for VMS.
 * sources-(final,media): Don't overwrite default sources.list.  The one shipped is the
   release that the ISO was built against, and has all the sections (-security, -backports)
   needed in case of stable builds too after the following commit.

If switching back, due to a regression in live-build --parent-debian-installer-distribution
must also be set to 'daily'.